requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
audiosegment==0.23.0
Babel==2.15.0
bangla==0.0.2
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.34.123
botocore==1.34.123
bs4==0.0.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
catalogue==2.0.10
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
EbookLib==0.18
einops==0.8.0
encodec==0.1.1
epub==0.5.2
filelock==3.15.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
g2pkk==0.1.2
gate-api==4.70.0
gcloud==0.18.3
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
gTTS==2.5.1
gunicorn==22.0.0
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
inflect==7.3.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonify==0.5
jsonlines==1.2.0
jws==0.1.3
kiwisolver==1.4.5
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.6.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mobi==0.3.3
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
openai==1.33.0
packaging==24.1
pandas==1.5.3
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pynndescent==0.5.13
pyparsing==3.1.2
pypdfium2==4.30.0
pypinyin==0.51.0
Pyrebase==3.0.27
Pyrebase4==4.8.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-jwt==2.0.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
trainer==0.0.36
transformers==4.41.2
translate==3.6.1
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.19
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webrtcvad==2.0.10
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
